refactor: 💡 [DateTimePicker] code refactored
✅ Closes: #336
