+++ src/components/DateTimePicker/index.tsx
@@ -3,26 +3,13 @@ import { AdapterLuxon } from "@mui/x-date-pickers/AdapterLuxon";
 import { DateTimePicker as MuiDateTimePicker } from "@mui/x-date-pickers/DateTimePicker";
 import { LocalizationProvider } from "@mui/x-date-pickers/LocalizationProvider";
 
-import { IBase } from "../../types/Base";
+import { DateTimePickerType, viewType } from "../../types/DateTimePicker";
 
 export const DATA_CY_DEFAULT = "timePicker";
-export type DateTimePickerType = IBase & {
-  label: string;
-  value?: string;
-  onAccept?: (value: string | null) => void;
-  ampm?: boolean;
-  mobileView?: boolean;
-  format?: string;
-  timeView?: "hrs" | "min" | "sec";
-};
-
-type viewType =
-  | ["year", "month", "day", "hours"]
-  | ["year", "month", "day", "hours", "minutes"]
-  | ["year", "month", "day", "hours", "minutes", "seconds"];
 
 const DateTimePicker: FC<DateTimePickerType> = ({
   dataCy = DATA_CY_DEFAULT,
+  timeZone,
   label = "Date Time",
   value,
   onAccept,
@@ -35,7 +22,7 @@ const DateTimePicker: FC<DateTimePickerType> = ({
     return mobileView ? "Mobile" : undefined;
   }, [mobileView]);
 
-  const views: viewType = useMemo(() => {
+  const views: viewType[] = useMemo(() => {
     if (timeView === "hrs") {
       return ["year", "month", "day", "hours"];
     }
@@ -47,6 +34,7 @@ const DateTimePicker: FC<DateTimePickerType> = ({
   return (
     <LocalizationProvider dateAdapter={AdapterLuxon}>
       <MuiDateTimePicker
+        timezone={timeZone}
         desktopModeMediaQuery={desctopMode}
         value={value}
         onAccept={onAccept}
